12.08패치노트
메이트 참여 조건추가(3개 모임까지 참여가능)
메이트찾기 신고기능추가

+++ src/main/java/com/kosta/jupjup/dao/MateFindDAO.java
@@ -32,6 +32,6 @@ public interface MateFindDAO {
 	
 	public Integer reportChk(@Param("id")String id,@Param("no") Long no);
 	
-	public void reportUpdate(int no);
+	public void reportUpdate(Long no);
 	
 }
