+++ src/main/java/com/kosta/jupjup/vo/ReportVO.java
@@ -9,7 +9,7 @@ public class ReportVO {
 	private int report_type;
 	private String report_writer;
 	private String report_content;
-	private int mate_id;
+	private Long mate_id;
 	private String user_id;
 	private String report_user_id;
 	private String report_mate_id;
